#85: Clean up test names and move the enum code to the correct package
